# Optimise .fill() throughput